# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: South African School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubjectBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: unreadOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Send notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationBody"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: learnerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: feeId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: processPayment
      tags: [payments]
      summary: Process a payment (simulated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessPaymentBody"
      responses:
        "201":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/receipt:
    get:
      operationId: getPaymentReceipt
      tags: [payments]
      summary: Get payment receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentReceipt"
  /timetable:
    get:
      operationId: listTimetableEntries
      tags: [timetable]
      summary: List timetable entries
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: teacherId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Timetable entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableEntry"
    post:
      operationId: createTimetableEntry
      tags: [timetable]
      summary: Create timetable entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimetableEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
  /timetable/{id}:
    patch:
      operationId: updateTimetableEntry
      tags: [timetable]
      summary: Update timetable entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimetableEntryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimetableEntry"
    delete:
      operationId: deleteTimetableEntry
      tags: [timetable]
      summary: Delete timetable entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        grade:
          type: string
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - code
        - grade
        - createdAt
    CreateSubjectBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        grade:
          type: string
        teacherId:
          type: ["integer", "null"]
      required:
        - name
        - code
        - grade
    UpdateSubjectBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        code:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        teacherId:
          type: ["integer", "null"]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [general, academic, financial, attendance, urgent]
        targetRole:
          type: ["string", "null"]
        targetUserId:
          type: ["integer", "null"]
        isRead:
          type: boolean
        sentBy:
          type: ["integer", "null"]
        senderName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - message
        - type
        - isRead
        - createdAt
    CreateNotificationBody:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [general, academic, financial, attendance, urgent]
        targetRole:
          type: ["string", "null"]
        targetUserId:
          type: ["integer", "null"]
        sentBy:
          type: ["integer", "null"]
      required:
        - title
        - message
        - type
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        feeId:
          type: integer
        learnerId:
          type: integer
        learnerName:
          type: ["string", "null"]
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [eft, credit_card, debit_card, cash, direct_deposit]
        referenceNumber:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        processedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - feeId
        - learnerId
        - amount
        - paymentMethod
        - referenceNumber
        - status
        - createdAt
    ProcessPaymentBody:
      type: object
      properties:
        feeId:
          type: integer
        learnerId:
          type: integer
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [eft, credit_card, debit_card, cash, direct_deposit]
        cardNumber:
          type: ["string", "null"]
        cardName:
          type: ["string", "null"]
      required:
        - feeId
        - learnerId
        - amount
        - paymentMethod
    PaymentReceipt:
      type: object
      properties:
        paymentId:
          type: integer
        referenceNumber:
          type: string
        learnerName:
          type: string
        admissionNumber:
          type: string
        description:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
        processedAt:
          type: string
        schoolName:
          type: string
      required:
        - paymentId
        - referenceNumber
        - learnerName
        - admissionNumber
        - description
        - amount
        - paymentMethod
        - processedAt
        - schoolName
    TimetableEntry:
      type: object
      properties:
        id:
          type: integer
        classId:
          type: integer
        className:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday]
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - classId
        - subjectId
        - dayOfWeek
        - startTime
        - endTime
        - createdAt
    CreateTimetableEntryBody:
      type: object
      properties:
        classId:
          type: integer
        subjectId:
          type: integer
        teacherId:
          type: ["integer", "null"]
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday]
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: ["string", "null"]
      required:
        - classId
        - subjectId
        - dayOfWeek
        - startTime
        - endTime
    UpdateTimetableEntryBody:
      type: object
      properties:
        subjectId:
          type: ["integer", "null"]
        teacherId:
          type: ["integer", "null"]
        dayOfWeek:
          type: ["string", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        room:
          type: ["string", "null"]
...