# lib/api-zod/src/generated/types/updateSubjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * South African School Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSubjectBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  teacherId?: number | null;
}